IBX-2493: Redo UserGroup ChoiceLoader
